refactor: merged page.ts and page.server.ts in admin page
